Improvement/conversion flag (#120)
* Separated encrypt and decrypt request handlers.

* Removed logic that checks for request type from the request handling ecalls.

* Added hooks for tag and iv to the encrypt and decrypt request handlers.

* Updated comments.

* Added arguments to request_decoder to provide space for tags and ivs.

* Added code to pull tag and iv from the json struct and changed some LOG_ERRs to OG_WARNs because a missing field is no longer necessarily an error.

* Added iv and tag as usable JSON fields and removed an unnecessary test case.

* Fixed bug that would prevent iv and tag from being encoded in the response message.

* Added tags and IVs to test cases.

* Stated adding tests for tag/iv parsing.

* Added additional check before freeing charbufs so it will work safely with uninitialized charbufs.

* Added parsing of cipher name from the request.

* Added cipher field to the json request structures.

* Added cipher_name argument to the request handling functions.

* Further integration of cipher structure and types.

* Changed request handler to derive the cipher from the request.

* Finished updating the request handler to properly handle any cipher.

* Added additional error handling.

* Adding support for AES-GCM in requests.

* Renamed functions to avoid collisions.

* Modified table tests to not segfault if a file is missing.

* Small updates to ensure tags and IVs work as expected, as well as unit test updates.

* Added extra error handling and cleaned up tests.

* Added checks for non-null key IDs.

* Added tests for null key IDs.

* Added check for missing key ID to the decrypt handler.

* Added tests for a missing or invalid cipher name.

* Added tests for missing input data to the request handlers.

* Fixed comments.

* Updated comments to reflect new arguments.

* Updated comments.

* Updated comments.

* Fixed comments.

* Removed unused file.

* Fixed bug that caused the tag and IV to not be included in the response to encrypt requests, and also fixed associated tests to test that correctly.

* Removed unnecessary length check in free_charbuf.

* Moved allocations closer to use in the AES GCM encryptor.

* Moved allocations closer to use in the AES GCM decryptor.

* Moved allocations to right before they're needed.

* Moved initialization (but not allocation) of return values to the start of the functions.

* Defined cipher_data_t structure.

* First cut at adding a cipher_data structure and incorporating it.

* Updated comments.

* Updated comments.

* Updated comments.

* Cleaned up error handling in the encrypt request handler.

* Some (but not all) test fixes.

* Removed early-out error in the JSON parser that ensured signed requests always fail. The (temporary) logic for that should be in validate_signature, rather than the parser.

* Started integration signature code with the new request fields.

* Added additional arguments to signature validation function.

* Finished initial integration of the new request fields with signature validation. Some tests are currently intentionally broken until the signature validation code is fully implemented.

* Removed signature validation from parsing code.

* Allowed request signing to succeed even though it's not fully implemented yet.

* Started moving signature verification into the enclave.

* Moved base64 decoding into the JSON parser.

* Removed dead code.

* Moved IV base-64 decoding into the decoder.

* Moved tag base-64 decoding into the decoder.

* Moved last of the base-64 decoding into the decoder.

* Added spots for the signature and certificate to the request handlers and moved base-64 encoding into the message encoder.

* Added additional arguments to request serialization and signature validation.

* Added serialization of additional request data.

* Updated comments.

* Removed unused code.

* Updated comments.

* Cleaned up serialization to always require tag and iv on decrypt requests.

* Starting to add tests for signed requests.

* Initial tests of the signed request functionality. Not fully functional yet.

* Continuing work on tests for signed requests.

* Changed sealing policy to MRENCLAVE.

* Added code to import CA keys in the tests.

* Previous change broke unit testing. Changes in this commit fix the unit testing by moving the sealing of sealed files needed by unit testing to the test enclave.

* Changed certificate format conversion to use a helper pointer so the original would still be available.

* Added additional test case.

* Update include/enclave_request_signing.h


* Removed unnecessary test file.

* Changed validate signature to return int instead of bool.

* Removed dead code.

* Changed verify_cert to return int instead of bool.

* Added -Wconversion flag to the Makefile and fixed warnings in src/util/enclave_request_signing.

* Checked all conversions in src/cipher/pelz_aes_gcm.c

* Checked all type conversions in src/util/pelz_socket.c.

* Changed table_id to use a size_t for the index argument, which is both semantically correct and fixes the conversion issue in fifo_thread.c

* Check conversions in src/util/secure_socket_thread.c

* Fixed slightly-off (but would still have worked) type conversion.

* Checked all type conversions in src/util/pipe_io.c

* Checked all type conversions in src/util/pelz_uri_helpers.c

* Changed functions to take size_t length arguments.

* Cleaned up types and conversions in cmd_interface files.

* Removed no-longer-necessary type conversions from src/pelz/main.c

* Cleaned up type conversions in common_table files.

* Cleaned up a variety of type conversions related to the tables.

* Cleaned up type conversions in src/cipher/pelz_aes_keywrap_3394nopad.c

* Finished cleaning up type conversions in pelz.

* Added missing include directive.

* Cleaned up type conversions in the unit tests.

* Fixed a type conversion I missed before.

* Fixed but introduced earlier in this branch in the last keywrap handling block.

* Removed unnecessary type conversions.

* Cleaned new warnings introduced by merging with upstream.

* Updated kmyth to latest main.

* Checked calloc returns.

---------

+++ include/cmd_interface.h
@@ -1,6 +1,8 @@
 #ifndef _CMD_INTERFACE_PELZ_H_
 #define _CMD_INTERFACE_PELZ_H_
 
+#include <stddef.h>
+
 /*
  *  0    seal <path>               Seals a file in .nkl format (pelz-service is not involved)
  *  1    exit                      Terminates running pelz-service
@@ -76,7 +78,7 @@ typedef enum
  *
  * @returns 0 on success, 1 on error
  */
-  int msg_arg(char *pipe, int pipe_len, int cmd, char *arg, int arg_len);
+  int msg_arg(char *pipe, size_t pipe_len, int cmd, char *arg, size_t arg_len);
 
 /**
  * @brief Creates and sends message then listens for response on the pipe provided.
@@ -92,7 +94,7 @@ typedef enum
  *
  * @returns 0 on success, 1 on error
  */
-  int msg_two_arg(char *pipe, int pipe_len, int cmd, char *arg, int arg_len, char *arg2, int arg2_len);
+  int msg_two_arg(char *pipe, size_t pipe_len, int cmd, char *arg, size_t arg_len, char *arg2, size_t arg2_len);
 
 /**
  * @brief Creates and sends message then listens for a list of responses on the pipe provided.
@@ -103,5 +105,5 @@ typedef enum
  *
  * @returns 0 on success, 1 on error
  */
-  int msg_list(char *pipe, int pipe_len, int cmd);
+  int msg_list(char *pipe, size_t pipe_len, int cmd);
 #endif
